use dma for framebuffer memory in st7586fb.c
This fixes an issue when writing to the framebuffer using mmap. Since Linux
memory is cached, we were getting artifacts displaying on the screen when
writing to the memory mapped file because the deferred io function was being
called long before the kernel memory was actually updated. DMA memory is not
cached, so this eliminates the problem.