cleanup of legoev3_ports and legoev3_uart
This includes major refactoring of ev3_input_port and ev3_output_port to make
it more clear when whether we are refering to the driver data or the parent
port device.

Also separate legoev3_port_driver from legoev3_port_device_driver. This
simplifies the implemetation of each. A legoev3_port(_driver) refers to
the input or output port itself whereas a legoev3_port_device(_driver)
refers to something that is pluged into a port.

Also added a port_name attribute and PORT env property to legoev3_port_device
so that it is easy to identify which port a device is plugged into without
having to dissect the device name or look at the parent device.

In the legoev3_uart module, removed registering an unneeded legoev3_port_device
and fixed a few issues with compiling on other platforms.

And finally, cleaned up a bunch of whitespace and comments.