More versitle custom scaling in lego-sensor class.
The existing method of scaling by calling a callback function each time a value
was updated resulted in much duplication of code. It also did not allow for the
scaled value to be larger than the raw value data type.

Scaling is now only performed when reading the value* attributes, which will
improve performace as well.

Had to remove debounce from wedo tilt sensor. A more generic debounce that can
be used by all sensors needs to be implemented instead. This should be done in
conjunction with adding sysfs_notify calls.

Also added PROXY mode to wedo motion sensor.