Remove set_mode function in da8xx musb driver.
This function sets USB0OTGMODE in CFGCHIP2. Since we are using peripheral-only
mode, this sets USB0OTGMODE to 2 which basically forces VBUS to always appear
high as if something is plugged in. This causes the musb driver to not be able
to detect disconnect events.

Since this behavior is not wanted/needed and the da8xx_musb_set_mode doesn't do
anything else, we just remove the whole function.

Issue ev3dev/ev3dev#244