Revert "Enable acls (needed by systemd)"
This reverts commit 559bc8a1254ccb6bb76fde4db0638bec2ea972be.

Several websites mention that using acls reduces performance and that it does
not make sense to turn it on system wide - although I could not find any hard
data to substantiate this. Nevertheless, we have been running fine without it
turned on so far, so we will leave them off for now.