Properly enable FIQ
The channel lookup code was buggy and it was just by chance that it
enabled the FIQ for us when UART1 was disabled and re-enabled if it
was selected to be a console tty.

The proper way to enable it is to do so explicitly after we have assigned
a FIQ handler. Otherwise, if we get an interrupt before the handler is
assigned, the processor will lock up.

Issue ev3dev/ev3dev#47