ARM: da8xx: Use PORT_DA830 UART type
This changes da8xx boards to use PORT_DA830 as the UART type. This fixes
a couple things. AFE capability was not propertly detected. By specifing
a fixed type, AFE capability is manually specified. Also, the UART reset
using the SoC-specific UART power management register is now handed in the
UART driver rather than in the mach code.

The check for PORT_AR7 was a nop since no mach-davinci boards specify that
port, so it seems that it can be safely removed.