remoteproc: assign device tree node to virtio device
This adds a device tree lookup to assign a device tree node to a virtio
device registered by a remoteproc device. The lookup expects a child
node of the remoteproc node named "vdevs". Each virtio device is a child
of this vdevs nodes. The "reg" property is used to match the notifyid
of the virtio devices.

Signed-off-by: David Lechner <david@lechnology.com>